+++ docs/CHANGELOG.md
@@ -93,6 +93,9 @@
   * `Force Touch GUI Button Behavior` To `Force Touch UI Button Behavior`
   * `Remove Forced GUI Lag (Can Break Joining Servers)` To `Remove Forced UI Lag (Can Break Joining Servers)`
   * `Hide Block Outline When GUI Is Hidden` To `Hide Block Outline When UI Is Hidden`
+  * `Fix Camera Functionality` To `Add Camera Functionality`
+  * `Fix Camera Rendering` To `Enable Camera Rendering`
+  * `Fix Camera Legs` To `Render Camera Legs`
 * Add Milk Buckets
   * Included In The `Add Buckets` Feature Flag
 * Removed `Property Scale Animated Textures` Feature Flag